MAINT: move expr submodule from utils to top-level
